+++ lanterns/src/components/DedicationToken/ded.css
@@ -1,7 +1,7 @@
 .tokenHand {
     
     text-align: center;
-    padding-top: 50px;
+    padding: 50px 10px 10px 10px;
     
    
 }
