+++ lanterns/src/pictures/3_pairs.jpg
